% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos_from_df.R
\name{plot_circos_from_df}
\alias{plot_circos_from_df}
\title{plot_circos_from_df}
\usage{
plot_circos_from_df(df)
}
\arguments{
\item{infile}{input file}
}
\value{

}
\description{
Plot a circos from a tibble
}
